pkg/ifuzz/powerpc: use GenerateInt()
GenerateInt() generates sort of random numbers for instruction fuzzer
with focus on corner cases, let's use it for POWERPC too.

Since we want memory access instruction to try these addresses, we preload
generated values in GPRs used by just generated instruction.
This in turn requires Insn::Encode() access for the instruction map to
encode load instructions so this moves ld64() from the generator to
insnSetMap and adds Insn::insnMap. This adds enc() to encode just
the instruction without any randomization.

Since EncodeParam() is not used by anything but Encode(), this open codes
it.

Signed-off-by: Alexey Kardashevskiy <aik@linux.ibm.com>