vm/qemu: dump linux'es printk buffer
This dumps the printk buffer using QEMU QMP. This is useful for crashed
when ssh connection lost as the buffer may still have clues what went
wrong; often it is out-of-memory killer.

This uses the current upstream format which is log_bug (256KB..1MB) and
2 ring buffers with message descriptors and extended information.
This prints the minimum to rely from the existing ring buffer format
as little as possible.

Signed-off-by: Alexey Kardashevskiy <aik@linux.ibm.com>
---

Is there a better way to parse string with bytes to []byte?

Is there a good way to stop suppressing crashes as now there is one more
way of extracting log from the vm? Ideally I want only suppressed crashes
to dump the printk buffer.

Is there a good way to find offsets within structures by just using
the elf package?