executor/common_kvm_ppc64: add an instruction for explicit KVM exit
KVM_RUN runs the generate instructions in a VM until it exits. We want
reproducible behavior so X86 adds "hlt" in the end of the generated blob.

This uses a software debug breakpoint instruction which is some invalid
instruction defined in KVM, hence the KVM_REG_PPC_DEBUG_INST ioctl.

Signed-off-by: Alexey Kardashevskiy <aik@linux.ibm.com>