executor/common_kvm_ppc64: run with enabled MMU
This sets up a page table to map the text in order to exercise
more code paths in the KVM.

This defines a flag to control the MMU state. When enabled, this
creates a simple page table at the 64K offset and maps all the RAM.
The fuzzer code is placed right after the table.

Signed-off-by: Alexey Kardashevskiy <aik@linux.ibm.com>