executor/common_kvm_ppc64: run with enabled MMU
This sets up a page table to map the text in order to exercise
more code paths in the KVM.

This defines flags to control the MMU state. When enabled, this
creates a simple page table at the 64K offset and maps all the RAM.
The fuzzer code is placed right after the table.

The flags are:
IR - enables MMU for instruction fetches
DR - enables MMU for data loads/stores
PR - "problem state", i.e. userspace (implies DR and IR)
PID1 - initializes a process table for PID>0 (PID#0 is used by the VM OS
       normally)

Signed-off-by: Alexey Kardashevskiy <aik@linux.ibm.com>